(기능2) 경주횟수 테스트 작성
