+++ src/main/java/com/mission2/racingcar/CarRacing.java
@@ -4,6 +4,7 @@ package com.mission2.racingcar;
 
 public class CarRacing {
     static final int MAX_NAME_COUNT = 5;
+    static final int MAX_GAME_COUNT = 10;
 
     public static void main(String[] args) {
